[test] Moved progress bar tests to test_selenium.py
